requirements.txt
contourpy==1.3.1
cycler==0.12.1
docopt==0.6.2
et_xmlfile==2.0.0
fonttools==4.56.0
imageio==2.37.0
importlib_resources==6.5.2
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
networkx==3.4.2
nibabel==5.3.2
numpy==2.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas-flavor==0.6.0
patsy==1.0.1
pillow==11.1.0
pykwalify==1.8.0
pyparsing==3.2.1
pyradiomics==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.8.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SimpleITK==2.4.1
six==1.17.0
statsmodels==0.14.4
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.3.13
typing_extensions==4.12.2
tzdata==2025.1
xarray==2025.1.2
